+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed:
 - Window manager persistence moved to settings database
+- Routing persistence moved to settings database
 
 ## [0.8.255] - 2023-02-02
 ### Changed: